Convert to using LowLevelDriver
This updates the JavaScript backend to suppor the new LowLevel driver
interface.  The purpose of this interface is to provide a more reusable
set of tools for when compiling to an imperative-like languages, such as
JavaScript.  The goal is to get more reuse across backends, and to avoid
continuablly reinventing the wheel.